cambiado permiso de mostrar gastos a eliminar gasto

+++ src/main/java/com/jobits/pos/ui/gastos/presenter/GastosViewPresenter.java
@@ -7,6 +7,7 @@ package com.jobits.pos.ui.gastos.presenter;
 
 import com.jgoodies.common.collect.ArrayListModel;
 import com.jobits.pos.controller.gasto.GastoOperacionController;
+import com.jobits.pos.controller.login.LogInController;
 import com.jobits.pos.domain.models.GastoVenta;
 import com.jobits.pos.domain.models.Venta;
 import com.jobits.pos.exceptions.ValidatingException;
@@ -113,11 +114,13 @@ public class GastosViewPresenter extends AbstractViewPresenter<GastosViewModel> 
     }
 
     private void onEliminarClick() {
-        if (getBean().getGasto_venta_seleccionado() == null) {
-            JOptionPane.showMessageDialog(Application.getInstance().getMainWindow(), "Seleccione un gasto primero");
-        } else {
-            service.removeGasto(getBean().getGasto_venta_seleccionado());
-            refreshState();
+        if (new LogInController().constructoAuthorizationView(null, R.NivelAcceso.ECONOMICO)) {
+            if (getBean().getGasto_venta_seleccionado() == null) {
+                JOptionPane.showMessageDialog(Application.getInstance().getMainWindow(), "Seleccione un gasto primero");
+            } else {
+                service.removeGasto(getBean().getGasto_venta_seleccionado());
+                refreshState();
+            }
         }
     }
